+++ CHANGELOG.md
@@ -1,8 +1,20 @@
-# 1.0.0-beta.2 (2020-11-30)
+# 1.0.0
+
+-   Added: make it **Gatsby v4** compatible
+
+# 1.0.0-beta.4
+
+-   Added: make it **Gatsby v3** compatible
+
+# 1.0.0-beta.3
+
+-   Removed: "@dontInfer" from node
+
+# 1.0.0-beta.2
 
 -   Added: "transform" option
 -   Added: "debug" option
 
-# 1.0.0-beta.1 (2020-05-20)
+# 1.0.0-beta.1
 
 -   Added: Initial commit